+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Updated/Enhanced Rules and Bug Fixes
  - #629: **ClosureAsLastMethodParameter**: Add *ignoreCallsToMethodNames* property.
  - #644: **NoTabCharacter**: Return actual line number containing (first) tab character in violation.
  - #646: **IllegalStringRule**: Return actual line number containing (first) matching string in violation.
+ - #638: **ClassStartsWithBlankLine**: Fix false positive if interface/class declaration takes more than 1 line.
 
 Build and Infrastructure
  - #610: Allow “CodeNarc.” prefix in @SuppressWarnings. ([Vladimir Orany](https://github.com/musketyr))